fix: blocking backup
This commit is supposed to fix proxy /api/v1/backup
not returning issue.

Closes #6.
